refactor: Moved all of the existing FastAPI code into server module
