+++ main.py
@@ -1,0 +1,3 @@
+import click
+
+
